requirements.txt
discord.py==1.3.1
SQLAlchemy==1.3.13
PyYAML==5.1.2